@@ -1,9 +1,10 @@
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
@@ -14,8 +15,8 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
     function step(op) {
         if (f) throw new TypeError("Generator is already executing.");
         while (_) try {
-            if (f = 1, y && (t = y[op[0] & 2 ? "return" : op[0] ? "throw" : "next"]) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [0, t.value];
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
             switch (op[0]) {
                 case 0: case 1: t = op; break;
                 case 4: _.label++; return { value: op[1], done: false };
@@ -34,18 +35,24 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
         if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
     }
 };
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
 Object.defineProperty(exports, "__esModule", { value: true });
-var _loadLibtimidity = require("./libtimidity");
-var Libtimidity = _loadLibtimidity();
-var version = Libtimidity.ccall("mid_get_version", "number", [], []);
-console.log("Libtimidity version " + version);
+exports.playerFromMIDIBuffer = void 0;
+var libtimidity_1 = __importDefault(require("./libtimidity"));
+var LibtimidityPromise = (0, libtimidity_1.default)();
+LibtimidityPromise.then(function (libtimidity) {
+    var version = libtimidity.ccall("mid_get_version", "number", [], []);
+    console.log("Libtimidity version ".concat(version));
+});
 function invariant(condition, msg) {
     if (!condition) {
-        throw new Error("Invariant violation: " + msg);
+        throw new Error("Invariant violation: ".concat(msg));
     }
 }
 var Player = /** @class */ (function () {
-    function Player(songPtr, bytesPerSample, channels, bufferSize, node, audioContext) {
+    function Player(songPtr, bytesPerSample, channels, bufferSize, node, audioContext, libtimidity) {
         var _this = this;
         this._listeners = [];
         this._destroyed = false;
@@ -75,7 +82,7 @@ var Player = /** @class */ (function () {
                 }
             }
             else {
-                invariant(false, _this._channels + " is not a supported number of channels.");
+                invariant(false, "".concat(_this._channels, " is not a supported number of channels."));
             }
             if (!_this._paused) {
                 if (sampleCount == 0) {
@@ -94,12 +101,13 @@ var Player = /** @class */ (function () {
         this._bytesPerSample = bytesPerSample;
         this._channels = channels;
         this._bufferSize = bufferSize;
-        this._bufferPtr = Libtimidity._malloc(bufferSize * this._bytesPerSample);
+        this._bufferPtr = libtimidity._malloc(bufferSize * this._bytesPerSample);
         this._node = node;
         this._array = new Int16Array(this._bufferSize * this._channels);
         this._audioContext = audioContext;
+        this._libtimidity = libtimidity;
         node.onaudioprocess = this._handleAudioProcess;
-        Libtimidity.ccall("mid_song_start", null, ["number"], [this._songPtr]);
+        libtimidity.ccall("mid_song_start", null, ["number"], [this._songPtr]);
     }
     Player.prototype.pause = function () {
         var _this = this;
@@ -117,12 +125,12 @@ var Player = /** @class */ (function () {
     Player.prototype.destroy = function () {
         invariant(!this._destroyed, "destroy() called after destroy()");
         this._listeners.splice(0, this._listeners.length);
-        Libtimidity.ccall("mid_song_free", null, ["number"], [this._songPtr]);
+        this._libtimidity.ccall("mid_song_free", null, ["number"], [this._songPtr]);
         this._songPtr = 0;
         this._destroyed = true;
         this._audioContext.close();
         this._node.disconnect();
-        delete this._node.onaudioprocess;
+        this._node.onaudioprocess = function () { };
     };
     /**
      * Seeks to the desired time in seconds
@@ -130,7 +138,7 @@ var Player = /** @class */ (function () {
     Player.prototype.seek = function (sec) {
         invariant(!this._destroyed, "seek() called after destroy()");
         var msec = Math.floor(sec * 1000);
-        Libtimidity.ccall("mid_song_seek", null, ["number", "number"], [this._songPtr, msec]);
+        this._libtimidity.ccall("mid_song_seek", null, ["number", "number"], [this._songPtr, msec]);
         this._notify();
     };
     /**
@@ -138,14 +146,14 @@ var Player = /** @class */ (function () {
      */
     Player.prototype.getTime = function () {
         invariant(!this._destroyed, "getTime() called after destroy()");
-        return Libtimidity.ccall("mid_song_get_time", "number", ["number"], [this._songPtr]) / 1000;
+        return this._libtimidity.ccall("mid_song_get_time", "number", ["number"], [this._songPtr]) / 1000;
     };
     /**
      * Returns duration of song in seconds.
      */
     Player.prototype.getDuration = function () {
         invariant(!this._destroyed, "getTotalTime() called after destroy()");
-        return Libtimidity.ccall("mid_song_get_total_time", "number", ["number"], [this._songPtr]) / 1000;
+        return this._libtimidity.ccall("mid_song_get_total_time", "number", ["number"], [this._songPtr]) / 1000;
     };
     Player.prototype.isPaused = function () {
         return this._paused;
@@ -159,12 +167,12 @@ var Player = /** @class */ (function () {
         this._listeners.splice(this._listeners.indexOf(fn), 1);
     };
     Player.prototype._renderToArray = function () {
-        var byteCount = Libtimidity.ccall("mid_song_read_wave", "number", ["number", "number", "number"], [this._songPtr, this._bufferPtr, this._bufferSize * this._bytesPerSample]);
+        var byteCount = this._libtimidity.ccall("mid_song_read_wave", "number", ["number", "number", "number"], [this._songPtr, this._bufferPtr, this._bufferSize * this._bytesPerSample]);
         var sampleCount = byteCount / this._bytesPerSample;
         if (sampleCount !== 0) {
             // Only bother rendering if we have content.
-            // TODO(joshuan): We assume s16! Extend this.
-            this._array.set(Libtimidity.HEAP16.subarray(this._bufferPtr / 2, (this._bufferPtr + byteCount) / 2));
+            // TODO(jocelyn): We assume s16! Extend this.
+            this._array.set(this._libtimidity.HEAP16.subarray(this._bufferPtr / 2, (this._bufferPtr + byteCount) / 2));
         }
         return sampleCount;
     };
@@ -183,42 +191,42 @@ function _getDefaultPreferences() {
         bufferSize: 8192,
     };
 }
-function _loadSong(midiData, prefs) {
-    var optionsPtr = Libtimidity.ccall("mid_alloc_options", "number", ["number", "number", "number", "number"], [prefs.rate, prefs.formatCode, prefs.channels, prefs.bufferSize]);
+function _loadSong(midiData, prefs, libtimidity) {
+    var optionsPtr = libtimidity.ccall("mid_alloc_options", "number", ["number", "number", "number", "number"], [prefs.rate, prefs.formatCode, prefs.channels, prefs.bufferSize]);
     // Create a stream
-    var midiBufferPtr = Libtimidity._malloc(midiData.byteLength);
-    Libtimidity.HEAPU8.set(new Uint8Array(midiData), midiBufferPtr);
-    var iStreamPtr = Libtimidity.ccall("mid_istream_open_mem", "number", ["number", "number"], [midiBufferPtr, midiData.byteLength]);
+    var midiBufferPtr = libtimidity._malloc(midiData.byteLength);
+    libtimidity.HEAPU8.set(new Uint8Array(midiData), midiBufferPtr);
+    var iStreamPtr = libtimidity.ccall("mid_istream_open_mem", "number", ["number", "number"], [midiBufferPtr, midiData.byteLength]);
     // Now, we can try to load the song itself
-    var songPtr = Libtimidity.ccall("mid_song_load", "number", ["number", "number"], [iStreamPtr, optionsPtr]);
+    var songPtr = libtimidity.ccall("mid_song_load", "number", ["number", "number"], [iStreamPtr, optionsPtr]);
     if (songPtr === 0) {
         // Something failed.
         // TODO: Get some kinda error info from the library?
-        Libtimidity.ccall("mid_istream_close", "number", ["number"], [iStreamPtr]);
-        Libtimidity._free(optionsPtr);
-        Libtimidity._free(midiBufferPtr);
+        libtimidity.ccall("mid_istream_close", "number", ["number"], [iStreamPtr]);
+        libtimidity._free(optionsPtr);
+        libtimidity._free(midiBufferPtr);
         throw new Error("Could not load that MIDI file.");
     }
     // We've got a song!
     // Clean up stuff we don't need any more
-    Libtimidity.ccall("mid_istream_close", "number", ["number"], [iStreamPtr]);
-    Libtimidity._free(optionsPtr);
-    Libtimidity._free(midiBufferPtr);
+    libtimidity.ccall("mid_istream_close", "number", ["number"], [iStreamPtr]);
+    libtimidity._free(optionsPtr);
+    libtimidity._free(midiBufferPtr);
     return songPtr;
 }
-function _loadPatchByName(name, patchUrlPrefix) {
+function _loadPatchByName(name, patchUrlPrefix, libtimidity) {
     return __awaiter(this, void 0, void 0, function () {
         var url, response, data, pathBits, basename, pathSoFar, i;
         return __generator(this, function (_a) {
             switch (_a.label) {
                 case 0:
                     console.log(name);
-                    url = "" + patchUrlPrefix + name;
+                    url = "".concat(patchUrlPrefix).concat(name);
                     return [4 /*yield*/, fetch(url)];
                 case 1:
                     response = _a.sent();
                     if (response.status !== 200) {
-                        throw new Error("Could not load " + url);
+                        throw new Error("Could not load ".concat(url));
                     }
                     return [4 /*yield*/, response.arrayBuffer()];
                 case 2:
@@ -228,14 +236,14 @@ function _loadPatchByName(name, patchUrlPrefix) {
                     pathSoFar = '/';
                     for (i = 0; i < pathBits.length; i++) {
                         try {
-                            Libtimidity.FS.mkdir(pathSoFar + pathBits[i]);
+                            libtimidity.FS.mkdir(pathSoFar + pathBits[i]);
                         }
                         catch (e) {
                             // do nothing
                         }
-                        pathSoFar += pathBits[i] + "/";
+                        pathSoFar += "".concat(pathBits[i], "/");
                     }
-                    Libtimidity.FS.writeFile(pathSoFar + basename, new Uint8Array(data), {
+                    libtimidity.FS.writeFile(pathSoFar + basename, new Uint8Array(data), {
                         encoding: 'binary',
                     });
                     return [2 /*return*/];
@@ -268,49 +276,52 @@ function playerFromMIDIBuffer(midiData, patchUrlPrefix) {
 exports.playerFromMIDIBuffer = playerFromMIDIBuffer;
 function _playerFromMIDIBuffer(midiData, patchUrlPrefix, audioContext) {
     return __awaiter(this, void 0, void 0, function () {
-        var prefs, node, config, configText, songPtr, missingPatchCount, newReqCount, player;
+        var prefs, libtimidity, node, config, configText, songPtr, missingPatchCount, newReqCount, player;
         return __generator(this, function (_a) {
             switch (_a.label) {
                 case 0:
                     prefs = _getDefaultPreferences();
+                    return [4 /*yield*/, LibtimidityPromise];
+                case 1:
+                    libtimidity = _a.sent();
                     node = audioContext.createScriptProcessor(prefs.bufferSize, 0, prefs.channels);
                     node.connect(audioContext.destination);
-                    if (!!_loaded) return [3 /*break*/, 3];
-                    return [4 /*yield*/, fetch(patchUrlPrefix + "timidity.cfg")];
-                case 1:
+                    if (!!_loaded) return [3 /*break*/, 4];
+                    return [4 /*yield*/, fetch("".concat(patchUrlPrefix, "timidity.cfg"))];
+                case 2:
                     config = _a.sent();
                     if (config.status >= 400) {
                         throw new Error("Could not get timidity.cfg");
                     }
                     return [4 /*yield*/, config.text()];
-                case 2:
-                    configText = _a.sent();
-                    Libtimidity.FS.writeFile("/timidity.cfg", configText);
-                    Libtimidity.ccall("mid_init", "number", ["string"], ["/timidity.cfg"]);
-                    _loaded = true;
-                    _a.label = 3;
                 case 3:
-                    songPtr = _loadSong(midiData, prefs);
-                    missingPatchCount = Libtimidity.ccall("mid_get_load_request_count", "number", ["number"], [songPtr]);
-                    if (!(missingPatchCount > 0)) return [3 /*break*/, 5];
+                    configText = _a.sent();
+                    libtimidity.FS.writeFile("/timidity.cfg", configText);
+                    libtimidity.ccall("mid_init", "number", ["string"], ["/timidity.cfg"]);
+                    _loaded = true;
+                    _a.label = 4;
+                case 4:
+                    songPtr = _loadSong(midiData, prefs, libtimidity);
+                    missingPatchCount = libtimidity.ccall("mid_get_load_request_count", "number", ["number"], [songPtr]);
+                    if (!(missingPatchCount > 0)) return [3 /*break*/, 6];
                     return [4 /*yield*/, Promise.all(// Continue once we load all the patches asynchronously.
                         Array(missingPatchCount).fill(null) // Create an array with reqCount nulls.
-                            .map(function (_, i) { return Libtimidity.ccall(// Get the names of the required patches
+                            .map(function (_, i) { return libtimidity.ccall(// Get the names of the required patches
                         "mid_get_load_request", "string", ["number", "number"], [songPtr, i]); })
-                            .map(function (patchName) { return _loadPatchByName(patchName, patchUrlPrefix); }))];
-                case 4:
+                            .map(function (patchName) { return _loadPatchByName(patchName, patchUrlPrefix, libtimidity); }))];
+                case 5:
                     _a.sent(); // Load 'em
                     // We need to try loading the song again, now that we've loaded the patches.
-                    Libtimidity.ccall("mid_song_free", null, ["number"], [songPtr]);
-                    songPtr = _loadSong(midiData, prefs);
-                    newReqCount = Libtimidity.ccall("mid_get_load_request_count", "number", ["number"], [songPtr]);
+                    libtimidity.ccall("mid_song_free", null, ["number"], [songPtr]);
+                    songPtr = _loadSong(midiData, prefs, libtimidity);
+                    newReqCount = libtimidity.ccall("mid_get_load_request_count", "number", ["number"], [songPtr]);
                     if (newReqCount !== 0) {
-                        Libtimidity.ccall("mid_song_free", null, ["number"], [songPtr]);
+                        libtimidity.ccall("mid_song_free", null, ["number"], [songPtr]);
                         throw new Error("Could not load all patches.");
                     }
-                    _a.label = 5;
-                case 5:
-                    player = new Player(songPtr, prefs.channels * (((prefs.formatCode & 0xFF) == 16) ? 2 : 1), prefs.channels, prefs.bufferSize, node, audioContext);
+                    _a.label = 6;
+                case 6:
+                    player = new Player(songPtr, prefs.channels * (((prefs.formatCode & 0xFF) == 16) ? 2 : 1), prefs.channels, prefs.bufferSize, node, audioContext, libtimidity);
                     return [2 /*return*/, player];
             }
         });
